types: Moved StorageBackend from core to types
