+++ core/src/backend.rs
@@ -1,6 +1,5 @@
 pub mod audio;
 pub mod navigator;
 pub mod render;
-pub mod storage;
 pub mod ui;
 pub mod video;
