+++ tests/tests/regression_tests.rs
@@ -3,10 +3,7 @@
 //! Trace output can be compared with correct output from the official Flash Player.
 
 use approx::assert_relative_eq;
-use ruffle_core::backend::{
-    navigator::{NullExecutor, NullNavigatorBackend},
-    storage::{MemoryStorageBackend, StorageBackend},
-};
+use ruffle_core::backend::navigator::{NullExecutor, NullNavigatorBackend};
 use ruffle_core::context::UpdateContext;
 use ruffle_core::events::MouseButton as RuffleMouseButton;
 use ruffle_core::external::Value as ExternalValue;
@@ -17,6 +14,7 @@ use ruffle_render_wgpu::target::TextureTarget;
 use ruffle_render_wgpu::wgpu;
 use ruffle_render_wgpu::WgpuRenderBackend;
 use ruffle_types::backend::log::LogBackend;
+use ruffle_types::backend::storage::{MemoryStorageBackend, StorageBackend};
 use ruffle_types::tag_utils::SwfMovie;
 use std::cell::RefCell;
 use std::collections::BTreeMap;
