+++ types/src/backend.rs
@@ -1,1 +1,2 @@
 pub mod log;
+pub mod storage;
